backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
qdrant-client==1.12.0
openai==1.52.0
sentence-transformers==3.1.1
torch==2.5.0
numpy==1.26.4
asyncio==3.4.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.35